feat: updated header styles
