+++ frontend/src/routes/(app)/home/[user]/+page.svelte
@@ -2,7 +2,6 @@
 	import Banner from '$lib/components/general/Banner.svelte';
 	import HomeCard from '$lib/components/home/HomeCard.svelte';
 	import HomeSearchBar from '$lib/components/home/HomeSearchBar.svelte';
-	import NewReportPopup from '$lib/components/popups/NewReportPopup.svelte';
 	import { EntrySort, EntryTypeFilter, ZenoService } from '$lib/zenoapi/index.js';
 	import { getContext } from 'svelte';
 
@@ -31,15 +30,7 @@
 	}
 </script>
 
-{#if showNewReport && data.user !== null}
-	<NewReportPopup
-		on:close={() => (showNewReport = false)}
-		user={data.user.name}
-		bind:showNewReport
-	/>
-{/if}
-
-<HomeSearchBar bind:typeFilter bind:searchText bind:sort bind:showNewReport myHub={true} />
+<HomeSearchBar bind:typeFilter bind:searchText bind:sort />
 <div class="grid h-full grid-cols-home content-start gap-5 overflow-y-auto">
 	{#each data.entries as entry ('id' in entry.entry ? entry.entry.id : 'uuid' in entry.entry ? entry.entry.uuid : '')}
 		<HomeCard entry={entry.entry} stats={entry.stats} user={data.user} />
